move tests to StacksafeTest + reset depth in AfterEach
